requirements.txt
# API
fastapi[standard]==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg[binary]==3.2.12

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Config
python-dotenv==1.0.1
pydantic-settings==2.6.1
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1
aiosqlite==0.20.0

# Linting & formatting
black==24.10.0
isort==5.13.2
flake8==7.1.1
